scrap/requirements.txt
google-cloud-storage
beautifulsoup4
lxml